+++ github3/api.py
@@ -139,8 +139,9 @@ def iter_gists(username=None, number=-1):
     return gh.iter_gists(username, number)
 
 
-def iter_repo_issues(owner, repository, filter='', state='', labels='',
-                     sort='', direction='', since='', number=-1):
+def iter_repo_issues(self, owner, repository, milestone=None, state=None,
+                     assignee=None, mentioned=None, labels=None, sort=None,
+                     direction=None, since=None, number=-1):
     """List issues on owner/repository. Only owner and repository are
     required.